+++ src/Phan/Language/Type/MixedType.php
@@ -58,4 +58,9 @@ final class MixedType extends NativeType
     {
         return true;  // It's possible.
     }
+
+    public function isValidNumericOperand() : bool
+    {
+        return true;
+    }
 }