+++ src/Phan/Language/Type/TrueType.php
@@ -54,4 +54,9 @@ final class TrueType extends ScalarType
         // This would be '1', which is probably not intended
         return Config::getValue('scalar_implicit_cast');
     }
+
+    public function isValidNumericOperand() : bool
+    {
+        return Config::getValue('scalar_implicit_cast');
+    }
 }